Rename files to keep git history
